Genomic Comparsion Formating Function

This group of functions allows you to creat a plot of -log10(P-values) of an association study by their genomic position, for example, the results of a GWAS or eQTL study. This function carries out the formatting necessary to generate a plot

gencom_format(assoc_data, chr_col, pos_col, p_col = NULL,
  log10p_col = NULL, ld_col = NULL)

Arguments

assoc_data
required. A dataframe that has columns containing the chromosome, physical position, and p-values or -log10(p-values) of the association, and can optionally have columns containing R2 information for LD in the region, or rsID numbers for the associated SNPs

chr_col
required. numeric. index of column in assoc_data containing chromosome information

pos_col
required. numeric. index of column in assoc_data containing genomic position information

p_col
required, if no log10p_col column specified. numeric. index of column in assoc_data containing p-values

log10p_col
required, if no p_col specified. numeric. index of column in assoc_data containing -log10(p-value)s

ld_col
optional. numeric. Required if you want to use the LD data in your data set in your plot, index of column in assoc_data containing LD information, e.g. R2 or D' values

rs_col
optional. numeric. Required if you want to calculated LD using 1000 genomes for analysis, index of column containing rsID numbers for SNPs

Examples

gencom_format(assoc_data = assoc_data, chr_col = 1, pos_col = 2, log10p_col = 3, ld_col = 4)

#> All inputs are go!

#> Formating association data...

#> Error in colnames(assoc_data)[chr_col] = "CHR": object 'assoc_data' not found